add an assert that could've saved my life
when pre-computing the jump table, i was placing the guard in every single square, in every possible direction, and running the simulation until she hit a wall
but I WAS PLACING HER ON WALLS AS WELL
i also marked the position with an X whenever she rotated, which was unnecessary.

so when there were two adjacent walls, and she spawned on one of them facing the other, and she would rotate...
she MARKED THE WALL UNDERNEATH HER WITH AN X
so the wall disappeared.

basically the mistake was placing her on walls in the first place
but this could have been spotted right away if i added an assert guard is not on wall whenever simulating a step